backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
# psycopg2-binary==2.9.9

# Security
bcrypt==4.1.2
python-dotenv==1.0.0
werkzeug==3.0.1

# Validation
marshmallow==3.20.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

requests

# AI / ML
ultralytics
opencv-python-headless
Pillow
